requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pytesseract==0.3.10
Werkzeug==2.2.3
zipp==3.13.0
